Added testing class, made changes to logistic_regression constructor if no args
